Add support for the Insignia NS-PCY5BMA Bluetooth Dongle
The Insignia NS-PCY5BMA Bluetooth Dongle is very similar to the IOGEAR
GBU521, except it uses the BCM20702B0 firmware, while the IOGEAR uses
the BCM20702A1 firmware. The Insignia Dongle also has different product
and vendor IDs and is incompatible with the A1 firmware. With the
proper firmware and device IDs, it works fine.